requirements.txt
cython
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
matplotlib==3.9.3
seaborn
termcolor==2.5.0
addict==2.4.0
yapf==0.43.0
PyYAML==6.0.2
Ipython==8.31.0

opencv-python== 4.10.0.84
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
timm==1.0.12
einops== 0.8.0
vit-pytorch==1.2.4
mxnet==1.8.0.post0
scikit-learn== 1.6.0

submitit
wandb
loralib==0.1.2

git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
git+https://github.com/facebookresearch/detectron2.git
